Move server implementation into subdir
